backend/requirements.txt
audioread==3.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
decorator==4.4.2
filelock==3.20.0
fsspec==2025.10.0
idna==3.11
jinja2==3.1.6
joblib==1.5.2
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markupsafe==3.0.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
platformdirs==4.5.0
pooch==1.8.2
pycparser==2.23
regex==2025.11.3
requests==2.32.5
resemblyzer==0.1.4
scikit-learn==1.7.2
scipy==1.15.3
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing==3.10.0.0
typing-extensions==4.15.0
urllib3==2.5.0
webrtcvad==2.0.10